software/contrib/andy/requirements.txt
multipledispatch
PySimpleGUI
"kivy[base]"
kivy_examples
